requirements.txt

requests
faiss-cpu
rank-bm25
sentence-transformers
transformers
peft
streamlit
pandas
tabulate
